requirements.txt
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
pytz==2021.3
screen==1.0.1
xlwt==1.3.0
